Do not report unavailable values as NaN
I am currently creating a Grafana dashboard for the ipmi exporter. And I find it incredibly difficult to deal with NaN values. For example, I want to calculate the average power consumption. But when a node is powered off, the power consumption is reported as NaN. Hence the entire result become none. My attempt to join the query with `ipmi_power_state` and to assume a default value instead of NaN while the node is powered off have been futile. In Prometheus, it is good practice to skip missing values. In this case, a simple `or` in the Prometheus query could be used to deal with the missing values. So, please consider to simply skip unavailable values rather than reporting them as NaN.